Revert "Merge branch 'main' of powerpook:IBAX-io/go-ibax"
This reverts commit 6494b9d5f89ae5a07d4f8b28e8093d37a68f8bc3, reversing
changes made to c075f7194e05b75bdfc3ddce6b9e21f8bad909d1.

+++ packages/storage/sqldb/ecosystem.go
@@ -40,7 +40,7 @@ type Combustion struct {
 type FeeModeInfo struct {
 	FeeModeDetail map[string]FeeModeFlag `json:"fee_mode_detail"`
 	Combustion    Combustion             `json:"combustion"`
-	FollowFuel    float64                `json:"follow_fuel"`
+	FollowFuel    int64                  `json:"follow_fuel"`
 }
 
 // TableName returns name of table
